compiler.h: remove Alpha/nds32/ARM-nonLinux blocks, create compiler_priv.h
- Remove obsolete architecture-specific code for Alpha, nds32, and ARM-nonLinux
- Move xserver-internal MMIO functions (xf86ReadMmio*, xf86WriteMmio*) to new compiler_priv.h
- Keep only public SDK symbols in compiler.h (mem_barrier, write_mem_barrier, barrier, inb/outb, MMIO macros, slowbcopy)
- Update fd_notify.h to define NotifyFdProcPtr (used by compiler_priv.h)
