+++ include/compiler.h
@@ -39,7 +39,7 @@
  * THE COPYRIGHT HOLDER(S) OR AUTHOR(S) BE LIABLE FOR ANY CLAIM, DAMAGES OR
  * OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
  * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- * OTHER DEALINGS IN THE SOFTWARE.
+ * PERFORMANCE OF THIS SOFTWARE.
  *
  * Except as contained in this notice, the name of the copyright holder(s)
  * and author(s) shall not be used in advertising or otherwise to promote
@@ -73,11 +73,6 @@
 #else
 #define mem_barrier() __asm__ __volatile__ ("lock; addl $0,0(%%esp)" : : : "memory")
 #endif
-
-#elif defined __alpha__
-
-#define mem_barrier() __asm__ __volatile__ ("mb" : : : "memory")
-#define write_mem_barrier() __asm__ __volatile__ ("wmb" : : : "memory")
 
 #elif defined __amd64__
 
@@ -138,78 +133,6 @@
 #endif
 
 #ifdef __GNUC__
-#if defined(__alpha__)
-
-#ifdef __linux__
-/* for Linux on Alpha, we use the LIBC _inx/_outx routines */
-/* note that the appropriate setup via "ioperm" needs to be done */
-/*  *before* any inx/outx is done. */
-
-extern _X_EXPORT void _outb(unsigned char val, unsigned long port);
-extern _X_EXPORT void _outw(unsigned short val, unsigned long port);
-extern _X_EXPORT void _outl(unsigned int val, unsigned long port);
-extern _X_EXPORT unsigned int _inb(unsigned long port);
-extern _X_EXPORT unsigned int _inw(unsigned long port);
-extern _X_EXPORT unsigned int _inl(unsigned long port);
-
-static inline void
-outb(unsigned long port, unsigned char val)
-{
-    _outb(val, port);
-}
-
-static inline void
-outw(unsigned long port, unsigned short val)
-{
-    _outw(val, port);
-}
-
-static inline void
-outl(unsigned long port, unsigned int val)
-{
-    _outl(val, port);
-}
-
-static inline unsigned int
-inb(unsigned long port)
-{
-    return _inb(port);
-}
-
-static inline unsigned int
-inw(unsigned long port)
-{
-    return _inw(port);
-}
-
-static inline unsigned int
-inl(unsigned long port)
-{
-    return _inl(port);
-}
-
-#endif                          /* __linux__ */
-
-#if (defined(__FreeBSD__) || defined(__OpenBSD__))
-
-/* for FreeBSD and OpenBSD on Alpha, we use the libio (resp. libalpha) */
-/*  inx/outx routines */
-/* note that the appropriate setup via "ioperm" needs to be done */
-/*  *before* any inx/outx is done. */
-
-extern _X_EXPORT void outb(unsigned int port, unsigned char val);
-extern _X_EXPORT void outw(unsigned int port, unsigned short val);
-extern _X_EXPORT void outl(unsigned int port, unsigned int val);
-extern _X_EXPORT unsigned char inb(unsigned int port);
-extern _X_EXPORT unsigned short inw(unsigned int port);
-extern _X_EXPORT unsigned int inl(unsigned int port);
-
-#endif                          /* (__FreeBSD__ || __OpenBSD__ ) */
-
-#if defined(__NetBSD__)
-#include <machine/pio.h>
-#endif                          /* __NetBSD__ */
-
 #elif defined(__amd64__) || defined(__i386__) || defined(__ia64__)
 
 #include <inttypes.h>
@@ -322,201 +245,6 @@ inl(unsigned long port)
     return ret;
 }
 
-static inline unsigned char
-xf86ReadMmio8(__volatile__ void *base, const unsigned long offset)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-    unsigned char ret;
-
-    __asm__ __volatile__("lduba [%1] %2, %0":"=r"(ret)
-                         :"r"(addr), "i"(ASI_PL));
-
-    return ret;
-}
-
-static inline unsigned short
-xf86ReadMmio16Be(__volatile__ void *base, const unsigned long offset)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-    unsigned short ret;
-
-    __asm__ __volatile__("lduh [%1], %0":"=r"(ret)
-                         :"r"(addr));
-
-    return ret;
-}
-
-static inline unsigned short
-xf86ReadMmio16Le(__volatile__ void *base, const unsigned long offset)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-    unsigned short ret;
-
-    __asm__ __volatile__("lduha [%1] %2, %0":"=r"(ret)
-                         :"r"(addr), "i"(ASI_PL));
-
-    return ret;
-}
-
-static inline unsigned int
-xf86ReadMmio32Be(__volatile__ void *base, const unsigned long offset)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-    unsigned int ret;
-
-    __asm__ __volatile__("ld [%1], %0":"=r"(ret)
-                         :"r"(addr));
-
-    return ret;
-}
-
-static inline unsigned int
-xf86ReadMmio32Le(__volatile__ void *base, const unsigned long offset)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-    unsigned int ret;
-
-    __asm__ __volatile__("lda [%1] %2, %0":"=r"(ret)
-                         :"r"(addr), "i"(ASI_PL));
-
-    return ret;
-}
-
-static inline void
-xf86WriteMmio8(__volatile__ void *base, const unsigned long offset,
-               const unsigned int val)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-
-    __asm__ __volatile__("stba %0, [%1] %2":    /* No outputs */
-                         :"r"(val), "r"(addr), "i"(ASI_PL));
-
-    barrier();
-}
-
-static inline void
-xf86WriteMmio16Be(__volatile__ void *base, const unsigned long offset,
-                  const unsigned int val)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-
-    __asm__ __volatile__("sth %0, [%1]":        /* No outputs */
-                         :"r"(val), "r"(addr));
-
-    barrier();
-}
-
-static inline void
-xf86WriteMmio16Le(__volatile__ void *base, const unsigned long offset,
-                  const unsigned int val)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-
-    __asm__ __volatile__("stha %0, [%1] %2":    /* No outputs */
-                         :"r"(val), "r"(addr), "i"(ASI_PL));
-
-    barrier();
-}
-
-static inline void
-xf86WriteMmio32Be(__volatile__ void *base, const unsigned long offset,
-                  const unsigned int val)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-
-    __asm__ __volatile__("st %0, [%1]": /* No outputs */
-                         :"r"(val), "r"(addr));
-
-    barrier();
-}
-
-static inline void
-xf86WriteMmio32Le(__volatile__ void *base, const unsigned long offset,
-                  const unsigned int val)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-
-    __asm__ __volatile__("sta %0, [%1] %2":     /* No outputs */
-                         :"r"(val), "r"(addr), "i"(ASI_PL));
-
-    barrier();
-}
-
-#elif defined(__arm32__) && !defined(__linux__)
-#define PORT_SIZE long
-
-extern _X_EXPORT unsigned int IOPortBase;      /* Memory mapped I/O port area */
-
-static inline void
-outb(unsigned PORT_SIZE port, unsigned char val)
-{
-    *(volatile unsigned char *) (((unsigned PORT_SIZE) (port)) + IOPortBase) =
-        val;
-}
-
-static inline void
-outw(unsigned PORT_SIZE port, unsigned short val)
-{
-    *(volatile unsigned short *) (((unsigned PORT_SIZE) (port)) + IOPortBase) =
-        val;
-}
-
-static inline void
-outl(unsigned PORT_SIZE port, unsigned int val)
-{
-    *(volatile unsigned int *) (((unsigned PORT_SIZE) (port)) + IOPortBase) =
-        val;
-}
-
-static inline unsigned int
-inb(unsigned PORT_SIZE port)
-{
-    return *(volatile unsigned char *) (((unsigned PORT_SIZE) (port)) +
-                                        IOPortBase);
-}
-
-static inline unsigned int
-inw(unsigned PORT_SIZE port)
-{
-    return *(volatile unsigned short *) (((unsigned PORT_SIZE) (port)) +
-                                         IOPortBase);
-}
-
-static inline unsigned int
-inl(unsigned PORT_SIZE port)
-{
-    return *(volatile unsigned int *) (((unsigned PORT_SIZE) (port)) +
-                                       IOPortBase);
-}
-
-#if defined(__mips__)
-#ifdef __linux__                    /* don't mess with other OSs */
-#if X_BYTE_ORDER == X_BIG_ENDIAN
-static inline unsigned int
-xf86ReadMmio32Be(__volatile__ void *base, const unsigned long offset)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-    unsigned int ret;
-
-    __asm__ __volatile__("lw %0, 0(%1)":"=r"(ret)
-                         :"r"(addr));
-
-    return ret;
-}
-
-static inline void
-xf86WriteMmio32Be(__volatile__ void *base, const unsigned long offset,
-                  const unsigned int val)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-
-    __asm__ __volatile__("sw %0, 0(%1)":        /* No outputs */
-                         :"r"(val), "r"(addr));
-}
-#endif
-#endif                          /* !__linux__ */
-#endif                          /* __mips__ */
-
 #elif defined(__powerpc__)
 
 #ifndef MAP_FAILED
@@ -524,111 +252,6 @@ xf86WriteMmio32Be(__volatile__ void *base, const unsigned long offset,
 #endif
 
 extern _X_EXPORT volatile unsigned char *ioBase;
-
-static inline unsigned char
-xf86ReadMmio8(__volatile__ void *base, const unsigned long offset)
-{
-    register unsigned char val;
-    __asm__ __volatile__("lbzx %0,%1,%2\n\t" "eieio":"=r"(val)
-                         :"b"(base), "r"(offset),
-                         "m"(*((volatile unsigned char *) base + offset)));
-    return val;
-}
-
-static inline unsigned short
-xf86ReadMmio16Be(__volatile__ void *base, const unsigned long offset)
-{
-    register unsigned short val;
-    __asm__ __volatile__("lhzx %0,%1,%2\n\t" "eieio":"=r"(val)
-                         :"b"(base), "r"(offset),
-                         "m"(*((volatile unsigned char *) base + offset)));
-    return val;
-}
-
-static inline unsigned short
-xf86ReadMmio16Le(__volatile__ void *base, const unsigned long offset)
-{
-    register unsigned short val;
-    __asm__ __volatile__("lhbrx %0,%1,%2\n\t" "eieio":"=r"(val)
-                         :"b"(base), "r"(offset),
-                         "m"(*((volatile unsigned char *) base + offset)));
-    return val;
-}
-
-static inline unsigned int
-xf86ReadMmio32Be(__volatile__ void *base, const unsigned long offset)
-{
-    register unsigned int val;
-    __asm__ __volatile__("lwzx %0,%1,%2\n\t" "eieio":"=r"(val)
-                         :"b"(base), "r"(offset),
-                         "m"(*((volatile unsigned char *) base + offset)));
-    return val;
-}
-
-static inline unsigned int
-xf86ReadMmio32Le(__volatile__ void *base, const unsigned long offset)
-{
-    register unsigned int val;
-    __asm__ __volatile__("lwbrx %0,%1,%2\n\t" "eieio":"=r"(val)
-                         :"b"(base), "r"(offset),
-                         "m"(*((volatile unsigned char *) base + offset)));
-    return val;
-}
-
-static inline void
-xf86WriteMmio8(__volatile__ void *base, const unsigned long offset,
-               const unsigned char val)
-{
-    __asm__
-        __volatile__("stbx %1,%2,%3\n\t":"=m"
-                     (*((volatile unsigned char *) base + offset))
-                     :"r"(val), "b"(base), "r"(offset));
-    eieio();
-}
-
-static inline void
-xf86WriteMmio16Le(__volatile__ void *base, const unsigned long offset,
-                  const unsigned short val)
-{
-    __asm__
-        __volatile__("sthbrx %1,%2,%3\n\t":"=m"
-                     (*((volatile unsigned char *) base + offset))
-                     :"r"(val), "b"(base), "r"(offset));
-    eieio();
-}
-
-static inline void
-xf86WriteMmio16Be(__volatile__ void *base, const unsigned long offset,
-                  const unsigned short val)
-{
-    __asm__
-        __volatile__("sthx %1,%2,%3\n\t":"=m"
-                     (*((volatile unsigned char *) base + offset))
-                     :"r"(val), "b"(base), "r"(offset));
-    eieio();
-}
-
-static inline void
-xf86WriteMmio32Le(__volatile__ void *base, const unsigned long offset,
-                  const unsigned int val)
-{
-    __asm__
-        __volatile__("stwbrx %1,%2,%3\n\t":"=m"
-                     (*((volatile unsigned char *) base + offset))
-                     :"r"(val), "b"(base), "r"(offset));
-    eieio();
-}
-
-static inline void
-xf86WriteMmio32Be(__volatile__ void *base, const unsigned long offset,
-                  const unsigned int val)
-{
-    __asm__
-        __volatile__("stwx %1,%2,%3\n\t":"=m"
-                     (*((volatile unsigned char *) base + offset))
-                     :"r"(val), "b"(base), "r"(offset));
-    eieio();
-}
 
 static inline void
 outb(unsigned short port, unsigned char value)
@@ -678,187 +301,6 @@ inl(unsigned short port)
     return xf86ReadMmio32Le((void *) ioBase, port);
 }
 
-#elif defined(__nds32__)
-
-/*
- * Assume all port access are aligned.  We need to revise this implementation
- * if there is unaligned port access.
- */
-
-#define PORT_SIZE long
-
-static inline unsigned char
-xf86ReadMmio8(__volatile__ void *base, const unsigned long offset)
-{
-    return *(volatile unsigned char *) ((unsigned char *) base + offset);
-}
-
-static inline void
-xf86WriteMmio8(__volatile__ void *base, const unsigned long offset,
-               const unsigned int val)
-{
-    *(volatile unsigned char *) ((unsigned char *) base + offset) = val;
-    barrier();
-}
-
-static inline unsigned short
-xf86ReadMmio16Swap(__volatile__ void *base, const unsigned long offset)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-    unsigned short ret;
-
-    __asm__ __volatile__("lhi %0, [%1];\n\t" "wsbh %0, %0;\n\t":"=r"(ret)
-                         :"r"(addr));
-
-    return ret;
-}
-
-static inline unsigned short
-xf86ReadMmio16(__volatile__ void *base, const unsigned long offset)
-{
-    return *(volatile unsigned short *) ((char *) base + offset);
-}
-
-static inline void
-xf86WriteMmio16Swap(__volatile__ void *base, const unsigned long offset,
-                    const unsigned int val)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-
-    __asm__ __volatile__("wsbh %0, %0;\n\t" "shi %0, [%1];\n\t":        /* No outputs */
-                         :"r"(val), "r"(addr));
-
-    barrier();
-}
-
-static inline void
-xf86WriteMmio16(__volatile__ void *base, const unsigned long offset,
-                const unsigned int val)
-{
-    *(volatile unsigned short *) ((unsigned char *) base + offset) = val;
-    barrier();
-}
-
-static inline unsigned int
-xf86ReadMmio32Swap(__volatile__ void *base, const unsigned long offset)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-    unsigned int ret;
-
-    __asm__ __volatile__("lwi %0, [%1];\n\t"
-                         "wsbh %0, %0;\n\t" "rotri %0, %0, 16;\n\t":"=r"(ret)
-                         :"r"(addr));
-
-    return ret;
-}
-
-static inline unsigned int
-xf86ReadMmio32(__volatile__ void *base, const unsigned long offset)
-{
-    return *(volatile unsigned int *) ((unsigned char *) base + offset);
-}
-
-static inline void
-xf86WriteMmio32Swap(__volatile__ void *base, const unsigned long offset,
-                    const unsigned int val)
-{
-    unsigned long addr = ((unsigned long) base) + offset;
-
-    __asm__ __volatile__("wsbh %0, %0;\n\t" "rotri %0, %0, 16;\n\t" "swi %0, [%1];\n\t":        /* No outputs */
-                         :"r"(val), "r"(addr));
-
-    barrier();
-}
-
-static inline void
-xf86WriteMmio32(__volatile__ void *base, const unsigned long offset,
-                const unsigned int val)
-{
-    *(volatile unsigned int *) ((unsigned char *) base + offset) = val;
-    barrier();
-}
-
-#if defined(NDS32_MMIO_SWAP)
-static inline void
-outb(unsigned PORT_SIZE port, unsigned char val)
-{
-    xf86WriteMmio8(IOPortBase, port, val);
-}
-
-static inline void
-outw(unsigned PORT_SIZE port, unsigned short val)
-{
-    xf86WriteMmio16Swap(IOPortBase, port, val);
-}
-
-static inline void
-outl(unsigned PORT_SIZE port, unsigned int val)
-{
-    xf86WriteMmio32Swap(IOPortBase, port, val);
-}
-
-static inline unsigned int
-inb(unsigned PORT_SIZE port)
-{
-    return xf86ReadMmio8(IOPortBase, port);
-}
-
-static inline unsigned int
-inw(unsigned PORT_SIZE port)
-{
-    return xf86ReadMmio16Swap(IOPortBase, port);
-}
-
-static inline unsigned int
-inl(unsigned PORT_SIZE port)
-{
-    return xf86ReadMmio32Swap(IOPortBase, port);
-}
-
-#else                           /* !NDS32_MMIO_SWAP */
-static inline void
-outb(unsigned PORT_SIZE port, unsigned char val)
-{
-    *(volatile unsigned char *) (((unsigned PORT_SIZE) (port))) = val;
-    barrier();
-}
-
-static inline void
-outw(unsigned PORT_SIZE port, unsigned short val)
-{
-    *(volatile unsigned short *) (((unsigned PORT_SIZE) (port))) = val;
-    barrier();
-}
-
-static inline void
-outl(unsigned PORT_SIZE port, unsigned int val)
-{
-    *(volatile unsigned int *) (((unsigned PORT_SIZE) (port))) = val;
-    barrier();
-}
-
-static inline unsigned int
-inb(unsigned PORT_SIZE port)
-{
-    return *(volatile unsigned char *) (((unsigned PORT_SIZE) (port)));
-}
-
-static inline unsigned int
-inw(unsigned PORT_SIZE port)
-{
-    return *(volatile unsigned short *) (((unsigned PORT_SIZE) (port)));
-}
-
-static inline unsigned int
-inl(unsigned PORT_SIZE port)
-{
-    return *(volatile unsigned int *) (((unsigned PORT_SIZE) (port)));
-}
-
-#endif                          /* NDS32_MMIO_SWAP */
-
-#endif                          /* arch madness */
-
 #else                           /* !GNUC */
 #if defined(__STDC__) && (__STDC__ == 1)
 #ifndef asm
@@ -873,121 +315,6 @@ inl(unsigned PORT_SIZE port)
 #define MMIO_IS_BE
 #endif
 
-#ifdef __alpha__
-static inline int
-xf86ReadMmio8(void *Base, unsigned long Offset)
-{
-    mem_barrier();
-    return *(CARD8 *) ((unsigned long) Base + (Offset));
-}
-
-static inline int
-xf86ReadMmio16(void *Base, unsigned long Offset)
-{
-    mem_barrier();
-    return *(CARD16 *) ((unsigned long) Base + (Offset));
-}
-
-static inline int
-xf86ReadMmio32(void *Base, unsigned long Offset)
-{
-    mem_barrier();
-    return *(CARD32 *) ((unsigned long) Base + (Offset));
-}
-
-static inline void
-xf86WriteMmio8(int Value, void *Base, unsigned long Offset)
-{
-    write_mem_barrier();
-    *(CARD8 *) ((unsigned long) Base + (Offset)) = Value;
-}
-
-static inline void
-xf86WriteMmio16(int Value, void *Base, unsigned long Offset)
-{
-    write_mem_barrier();
-    *(CARD16 *) ((unsigned long) Base + (Offset)) = Value;
-}
-
-static inline void
-xf86WriteMmio32(int Value, void *Base, unsigned long Offset)
-{
-    write_mem_barrier();
-    *(CARD32 *) ((unsigned long) Base + (Offset)) = Value;
-}
-
-extern _X_EXPORT void xf86SlowBCopyFromBus(unsigned char *, unsigned char *,
-                                           int);
-extern _X_EXPORT void xf86SlowBCopyToBus(unsigned char *, unsigned char *, int);
-
-/* Some macros to hide the system dependencies for MMIO accesses */
-/* Changed to kill noise generated by gcc's -Wcast-align */
-#define MMIO_IN8(base, offset) xf86ReadMmio8((base), (offset))
-#define MMIO_IN16(base, offset) xf86ReadMmio16((base), (offset))
-#define MMIO_IN32(base, offset) xf86ReadMmio32((base), (offset))
-
-#define MMIO_OUT8(base, offset, val) \
-    xf86WriteMmio8((CARD8)(val), (base), (offset))
-#define MMIO_OUT16(base, offset, val) \
-    xf86WriteMmio16((CARD16)(val), (base), (offset))
-#define MMIO_OUT32(base, offset, val) \
-    xf86WriteMmio32((CARD32)(val), (base), (offset))
-
-#elif defined(__powerpc__) || defined(__sparc__)
- /*
-  * we provide byteswapping and no byteswapping functions here
-  * with byteswapping as default,
-  * drivers that don't need byteswapping should define MMIO_IS_BE
-  */
-#define MMIO_IN8(base, offset) xf86ReadMmio8((base), (offset))
-#define MMIO_OUT8(base, offset, val) \
-    xf86WriteMmio8((base), (offset), (CARD8)(val))
-
-#if defined(MMIO_IS_BE)     /* No byteswapping */
-#define MMIO_IN16(base, offset) xf86ReadMmio16Be((base), (offset))
-#define MMIO_IN32(base, offset) xf86ReadMmio32Be((base), (offset))
-#define MMIO_OUT16(base, offset, val) \
-    xf86WriteMmio16Be((base), (offset), (CARD16)(val))
-#define MMIO_OUT32(base, offset, val) \
-    xf86WriteMmio32Be((base), (offset), (CARD32)(val))
-#else                           /* byteswapping is the default */
-#define MMIO_IN16(base, offset) xf86ReadMmio16Le((base), (offset))
-#define MMIO_IN32(base, offset) xf86ReadMmio32Le((base), (offset))
-#define MMIO_OUT16(base, offset, val) \
-     xf86WriteMmio16Le((base), (offset), (CARD16)(val))
-#define MMIO_OUT32(base, offset, val) \
-     xf86WriteMmio32Le((base), (offset), (CARD32)(val))
-#endif
-
-#elif defined(__nds32__)
- /*
-  * we provide byteswapping and no byteswapping functions here
-  * with no byteswapping as default; when endianness of CPU core
-  * and I/O devices don't match, byte swapping is necessary
-  * drivers that need byteswapping should define NDS32_MMIO_SWAP
-  */
-#define MMIO_IN8(base, offset) xf86ReadMmio8((base), (offset))
-#define MMIO_OUT8(base, offset, val) \
-    xf86WriteMmio8((base), (offset), (CARD8)(val))
-
-#if defined(NDS32_MMIO_SWAP)    /* byteswapping */
-#define MMIO_IN16(base, offset) xf86ReadMmio16Swap((base), (offset))
-#define MMIO_IN32(base, offset) xf86ReadMmio32Swap((base), (offset))
-#define MMIO_OUT16(base, offset, val) \
-    xf86WriteMmio16Swap((base), (offset), (CARD16)(val))
-#define MMIO_OUT32(base, offset, val) \
-    xf86WriteMmio32Swap((base), (offset), (CARD32)(val))
-#else                           /* no byteswapping is the default */
-#define MMIO_IN16(base, offset) xf86ReadMmio16((base), (offset))
-#define MMIO_IN32(base, offset) xf86ReadMmio32((base), (offset))
-#define MMIO_OUT16(base, offset, val) \
-     xf86WriteMmio16((base), (offset), (CARD16)(val))
-#define MMIO_OUT32(base, offset, val) \
-     xf86WriteMmio32((base), (offset), (CARD32)(val))
-#endif
-
-#else                           /* !__alpha__ && !__powerpc__ && !__sparc__ */
-
 #define MMIO_IN8(base, offset) \
 	*(volatile CARD8 *)(((CARD8*)(base)) + (offset))
 #define MMIO_IN16(base, offset) \
@@ -1001,20 +328,12 @@ extern _X_EXPORT void xf86SlowBCopyToBus(unsigned char *, unsigned char *, int);
 #define MMIO_OUT32(base, offset, val) \
 	*(volatile CARD32 *)(void *)(((CARD8*)(base)) + (offset)) = (val)
 
-#endif                          /* __alpha__ */
-
-/*
- * With Intel, the version in os-support/misc/SlowBcopy.s is used.
- * This avoids port I/O during the copy (which causes problems with
- * some hardware).
- */
-#ifdef __alpha__
-#define slowbcopy_tobus(src,dst,count) xf86SlowBCopyToBus((src),(dst),(count))
-#define slowbcopy_frombus(src,dst,count) xf86SlowBCopyFromBus((src),(dst),(count))
-#else                           /* __alpha__ */
 #define slowbcopy_tobus(src,dst,count) xf86SlowBcopy((src),(dst),(count))
 #define slowbcopy_frombus(src,dst,count) xf86SlowBcopy((src),(dst),(count))
-#endif                          /* __alpha__ */
+
+extern _X_EXPORT void xf86SlowBCopyFromBus(unsigned char *, unsigned char *,
+                                           int);
+extern _X_EXPORT void xf86SlowBCopyToBus(unsigned char *, unsigned char *, int);
 
 /* NOLINTEND(hicpp-no-assembler) */
